Moved logging of metadata
Moved logging of Accumulo metadata from tablet to its own logger.  This
allows logging of metadata to be turned on/off independantly from tablet
lifecycle events.
